requirements.txt
Flask
opencv-python-headless
requests
python-dotenv
